% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchDownload.R
\name{BatchDownload}
\alias{BatchDownload}
\title{BatchDownload}
\usage{
BatchDownload(namelist, GSMdownloadpath = ".")
}
\arguments{
\item{namelist}{A character list of GSE or GSM ID.}

\item{GSMdownloadpath}{A character value specifying the path to download the SRA files on users own Linux/Unix computer.}
}
\value{
A character string of commands to be run on users' computer.
}
\description{
Generate shell file to download the SRA of multiple GSE or GSM
}
\details{
This function returns a shell file that can download the SRA of multiple GSE or GSM. The shell file should be run on user's own computer with Linux/Unix system and wget installed.
The download process is to set up a bunch of directories under the path specified by users (GSMdownloadpath) and download the SRA files inside the paths.
Note: it is the users' job to make sure the GSE and GSM names are correct. Otherwise the function will run forever due to inner error control system.
Also, it could take a long time for the function to complete if there are lots of GSE series in namelist.
}
\examples{
BatchDownload("GSM1274542")
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}
