% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/database.R
\name{database}
\alias{database}
\title{List of species database}
\format{A data.frame with two columns. First column: name of Bioconductor database; Second column: corresponding species name.}
\source{
\url{http://www.bioconductor.org/}
}
\description{
The list of species annotation databases used in alias searching. All databases are from Bioconductor.
}

