% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEOsearch.R
\docType{package}
\name{GEOsearch}
\alias{GEOsearch}
\alias{GEOsearch-package}
\title{GEOsearch: Extendable Search Engine for Gene Expression Omnibus}
\description{
GEOsearch provides extended search function (i.e. search all aliases of a gene name) as well as other utilities.
}
\details{
GEOsearch is an extendable search engine for NCBI GEO (Gene Expression Omnibus). Instead of directly searching the term, GEOsearch can find all the gene names contained in the search term
and search all the alias of the gene names simultaneously in GEO database. GEOsearch also provides other functions such as summarizing common biology keywords in the search results.
}

