% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEOSearchTerm.R
\name{GEOSearchTerm}
\alias{GEOSearchTerm}
\title{GEOSearchTerm}
\usage{
GEOSearchTerm(termlist)
}
\arguments{
\item{termlist}{A character vector of terms to be searched in NCBI GEO. Typically the direct output of function "termalias"}
}
\value{
A data frame containing the search results returned from NCBI GEO. First column: GEO Series Accesion Number; Second column: Organism; Third column: Title; Fourth column: Type of experiment; Fifth column: Experiment Platform Sixth column: Number of Samples; Seventh column: URL for SRX download; Eighth column: Description Ninth column: Corresponding search term.
}
\description{
Perform searching in NCBI GEO database.
}
\details{
Search the terms one by one in NCBI GEO database and return an integrated table of search results. The returned results should contain exactly the same information as the results returned by directly searching in http://www.ncbi.nlm.nih.gov/geo/.
}
\examples{
GEOSearchTerm("Oct4 RNA-seq")
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

