% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GEOSearchUI.R
\name{GEOSearchUI}
\alias{GEOSearchUI}
\title{GEOSearchUI}
\usage{
GEOSearchUI()
}
\description{
Launch the GEOsearch graphical user interface
}
\details{
This function will automatically launch the GEOsearch user interface in a web browser.
The user interface provides many powerful functions (e.g. second-round search) which is not available by command line programming.
It also provides a much easier and more convenient way to perform all functions available in command lines.
The user interface can also be accessed by http://zhiji.shinyapps.io/GEOsearch. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
if(interactive()) {
   GEOSearchUI()
}
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

