% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KeyWordFreq.R
\name{KeyWordFreq}
\alias{KeyWordFreq}
\title{KeyWordFreq}
\usage{
KeyWordFreq(searchtable, category = c("celltype", "disease", "tissue"))
}
\arguments{
\item{searchtable}{The direct output of function "GEOsearchterm"}

\item{category}{A character vector specifying which category in the common biology keyword list to be used. Should be contain "celltype", "disease" or "tissue".}
}
\value{
A data.frame with the frequency of each common biology keyword. First column: keyword name; Second column: kyeword frequency; Third column: log fold change of the frequency; Fourth column: FDR of fisher test.
}
\description{
Frequencies of common biology keywords appearing in search results
}
\details{
This function calculates the frequencies of each common biology keyword appearing in the given search table. The list of common biology keywords is compiled from http://www.atcc.org/. The list contains three categories: cell types, diseases and tissues. Users can specify which category to be used. The function also returns log fold change and FDR of fisher test to check whether each keyword has significantly more appearance compared to base frequency. The base frequency is defined as the number of appearance of the key word in all samples (roughly 40000 samples) included in GEO database.
}
\examples{
KeyWordFreq(GEOSearchTerm("Oct4 RNA-seq"))
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

