% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/term.R
\name{term}
\alias{term}
\title{List of common biology keywords}
\format{A data.frame with three columns. First column: keyword name; Second column: category; Third column: number of appearance in GEO database.}
\source{
\url{http://www.atcc.org/}
}
\description{
The list of common biology keywords is compiled from http://www.atcc.org/. The list contains three categories: cell types, diseases and tissues.
}

