% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TermAlias.R
\name{TermAlias}
\alias{TermAlias}
\title{TermAlias}
\usage{
TermAlias(searchterm, allspecies = c("Human", "Mouse"), mincount = 5)
}
\arguments{
\item{searchterm}{A character value specifying the term being searched. Words should be separated by space.}

\item{allspecies}{A character vector specifying the species in which the alias of gene names will be searched for. Available species: Anopheles,Arabidopsis,Bovine,Worm,Canine,Fly,Zebrafish,E coli strain K12,E coli strain Sakai,Chicken,Human,Mouse,Rhesus,Malaria,Chimp,Rat,Yeast,Pig,Xenopus}

\item{mincount}{An integer value specifying the minimum number of appearance in the GEO database for the alias to be retained. If an alias has too few appearance it may not be of great interest.}
}
\value{
A character vector of combinatory results of alias.
}
\description{
Search Alias of a Term
}
\details{
This function first picks gene names from the searh term. It then searches alias for all the gene names contained in the search term. It next queries GEO and retain alias that appear frequently enough in GEO database. Finally it returns a combinatory results of retained alias.
}
\examples{
TermAlias("Oct4 RNA-seq")
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

